+++ src/textual/views/_window_view.py
@@ -13,7 +13,8 @@ from ..widgets import Static
 
 
 class WindowChange(Message):
-    pass
+    def can_replace(self, message: Message) -> bool:
+        return isinstance(message, WindowChange)
 
 
 class WindowView(View, layout=VerticalLayout):
@@ -24,7 +25,6 @@ class WindowView(View, layout=VerticalLayout):
         gutter: tuple[int, int] = (0, 1),
         name: str | None = None
     ) -> None:
-        self.gutter = gutter
         layout = VerticalLayout(gutter=gutter)
         self.widget = widget if isinstance(widget, Widget) else Static(widget)
         layout.add(self.widget)
@@ -38,6 +38,10 @@ class WindowView(View, layout=VerticalLayout):
         layout.add(self.widget)
         await self.refresh_layout()
         self.refresh(layout=True)
+        await self.emit(WindowChange(self))
+
+    async def message_update(self, message: UpdateMessage) -> None:
+        message.prevent_default()
         await self.emit(WindowChange(self))
 
     async def watch_virtual_size(self, size: Size) -> None: