+++ src/textual/widgets/_scroll_view.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from logging import PlaceHolder
 
 from rich.console import RenderableType
 from rich.style import StyleType
@@ -8,11 +7,9 @@ from rich.style import StyleType
 from .. import events
 from ..layouts.grid import GridLayout
 from ..message import Message
-from ..messages import UpdateMessage
+from ..messages import CursorMoveMessage
 from ..scrollbar import ScrollTo, ScrollBar
-from ..geometry import clamp, Offset, Size
-from ..page import Page
-from ..reactive import watch
+from ..geometry import clamp
 from ..view import View
 from ..widget import Widget
 
@@ -121,6 +118,21 @@ class ScrollView(View):
         self.target_x += self.size.width
         self.animate("x", self.target_x, speed=120, easing="out_cubic")
 
+    def scroll_in_to_view(self, line: int) -> None:
+        if line < self.y:
+            self.y = line
+        elif line >= self.y + self.size.height:
+            self.y = line - self.size.height + 1
+
+    def scroll_to_center(self, line: int) -> None:
+        self.target_y = line - self.size.height // 2
+        if abs(self.target_y - self.y) > 1:
+            # Animate if its more than 1
+            self.animate("y", self.target_y, easing="out_cubic")
+        else:
+            # Jump if its just one step
+            self.y = self.target_y
+
     async def on_mouse_scroll_up(self, event: events.MouseScrollUp) -> None:
         self.scroll_up()
 
@@ -191,3 +203,7 @@ class ScrollView(View):
             self.refresh()
         if self.layout.show_row("hscroll", virtual_size.width > self.size.width):
             self.refresh()
+
+    def message_cursor_move(self, message: CursorMoveMessage) -> None:
+        self.scroll_to_center(message.line)
+        message.stop()